New folder/requirements.txt
django
google-generativeai
python-dotenv
scikit-learn
pandas
numpy